Renaming library to libopal
