+++ include/libopal.h
